renamed /backend/route/index.js to app.js
